Move StoryPromo in common test
